Fix Bouffalo follow-up build and PDS issues
Update Bouffalo follow-up changes for PDS safety, OTA metadata parsing, logging format cleanup, generated contact sensor data model, and build target testdata.

+++ examples/lighting-app/bouffalolab/common/AppTask.cpp
@@ -163,7 +163,7 @@ void AppTask::AppTaskMain(void * pvParameter)
     resetCnt++;
     Internal::BflbConfig::WriteConfigValue(APP_REBOOT_RESET_COUNT_KEY, resetCnt);
     GetAppTask().mButtonPressedTime = System::SystemClock().GetMonotonicMilliseconds64().count();
-    ChipLogProgress(NotSpecified, "AppTaskMain %lld, resetCnt %ld", GetAppTask().mButtonPressedTime, resetCnt);
+    ChipLogProgress(NotSpecified, "AppTaskMain %llu, resetCnt %u", (unsigned long long) GetAppTask().mButtonPressedTime, (unsigned int) resetCnt);
 #endif
 
     GetAppTask().sTimer =
